Switch to mocha as the library test runner.
While it was convenient in some ways to pull in jest and use it for
the expect and mock dependencies, it was also causing some trouble
primarily this was because we are trying to emulate a part of jest.

Sidestep this issue by using mocha as the test runner. Make the
dependencies on pieces explicit and adapt the output in one case.