Rework .toHaveProperty() to correctly hander property presence.
In the case of no value being supplied as an additional argument
to the assertion opt to check property presence in the subject.
This cleans up a number of special cases and as a side effect
improves the clarity of the output.

The new approach does however open up an issue with properties
that are present but declared undefined in the subject. Unexpected
opts to make this equivalent to the property not being present, so
To deal with this here we add a special type for hasOwnProperty
comparisons and generate such within keyPathToSatisfySpec().