Pull a few ticks to clean up output on mismatched .toContainEqual().
Route all contain checks through an assertion in jest unexpected.
Accept more types than we would normally which means we can uphold
the jestExpect behaviour of nicer output on type mismatch.

While here also use alternations which makes the output on failure
serialise "to contain equal" as the name of the assertion.