Fix issues in .rejects.toThrow() chained assertion with an argument.
The code was was previously unpacking the error from the rejected
promise but the .toThrow() override, necessary because capturing
that error changes the active subject, was not building an assertion
function capable of consuming a .toThrow() argument and checking the
error against it. Make this happen correctly.

While here also prevent .resolves.toThrow() because such a construct
makes no sense.